Add edges between vertices
Edges represent the connection between two vertices. The `%Edge{}`
struct is directed, consisting of two `%Vertex{}`s: `point`, and `next`,
and the length between them.

Length is calculated via Pythagorean theorem.

For three edges, A -> B -> C, the angle b can be determined with
`calculate_angle/3`. Edges are cast to vectors `AB` and `BC`; the angle
`b` is the inverse cosine of the dot product of `AB` and `BC` over
`||AB||` * `||BC||` (conveniently, these are the lengths of the edges).