Remove dedicated RegularPolygon module
Regular polygons can be expressed entirely as general polygons. Keeping
track of the number of sides and rotation angle aren't as necessary as
the edges and vertices.

With the removal of `RegularPolygon`, a `gen_regular_polygon` function
is added to `Polygon` that will generate a regular polygon from its
number of sides, circumradius, intial rotation angle, and midpoint.

Tests are implemented for both regular polygons and general polygons in
`PolygonTest`.