Handle concave polygons in detection/resolution
On polygon creation, check whether it is convex or not and set it in the
struct.

For `collision?`, `resolution`, and `resolve_collision`, add additional
function heads that pattern match on whether one of the polygons is
concave. If one or either is concave, construct a convex hull using
`Vertex.graham_scan` and continue with the function call using the
convex hull. When a polygon is returned (e.g. `resolve_collision`), the
original polygon is translated and returned.